Add custom DAO, remove flag, extract job
